Update motion interpolation settings in portable_config/input.conf
Adjusted the values for `buffered-frames` and `concurrent-frames` in the `CTRL+m` command options for motion interpolation.
The new values aim to optimize performance and reduce stuttering or slower frame rates.
The `buffered-frames` value has been set to 8 and the `concurrent-frames` value has been set to 16.
These values can be further fine-tuned based on the number of logical processors in the CPU.