refactor: Modify motion estimation parameters in motion.vpy script
The `motion.vpy` script was modified to improve the accuracy of motion estimation. The `pel` parameter in the `core.mv.Super` function was changed to 4 for better accuracy. Additionally, the `blksize` parameter in the `core.mv.Analyse` function was changed to 64 for both forward and backward motion vectors. Lastly, the `mask` parameter in the `core.mv.FlowFPS` function was changed to 2 to adjust the mask for motion-compensated frame interpolation.