Update README and reconfigure keybinds and settings
Updated keybindings and settings in the README and .conf files for a clearer and more efficient user interface. Removed unnecessary details and rephrased explanations for better comprehension, and adjusted subtitles font and border size. Also updated and added different keybindings for better user experience, while renaming one of the scripts for better relevance.